The application programming interface.

Functions

SoftSol3D.LocalMaxSpeeds — Method

LocalMaxSpeeds(Q::Array{Float64,1}, mate::Material, cmaxloc::Array{Float64,1})

Returns the local maximum characteristic speeds along x, y, z at the state vector Q with the parameters of the argument.

SoftSol3D.Run — Method

Run(mate::Material, mesh::Mesh, src::Source)

Executes the main loop for the parameters specified in the argument. The numerical solution at the final time data.t is stored in data.Qp. This data vector of size $6N+13$ has the following structure:

${\bf Q} = \left(F_{11}, F_{12}, F_{13}, \dots , F_{33}, v_1, v_2, v_3, [S_1^\text{v}]_{11}, \dots , [S_N^\text{v}]_{33}\right)^\top$

where $F_{ij}$ are the deformation gradient components, $v_i$ are the velocity components, and $[S_\ell^\text{v}]_{ij}$ are the memory variables.

SoftSol3D.Slice — Method

x, u = Slice(dir::Int64, mesh::Mesh, data::Data)

Returns data data.Qp sliced along direction dir in output u. Spatial coordinates mesh.X in output x undergo the same operation.

SoftSol3D.compSol — Method

xth, uth, res = compSol(mate::Material, mesh::Mesh, src::Source)

Computes analytical solutions via FFT.

SoftSol3D.ConfigPar — Module

Defines the data types using Parameters.jl.

SoftSol3D.ConfigPar.initSource — Method

pts, coeffs = initSource(mate::Material, mesh::Mesh, src::Source)

Source configuration. Returns source points pts and corresponding body force coefficients coeffs.

SoftSol3D.ConfigPar.Material — Type

Material parameters.

SoftSol3D.ConfigPar.Mesh — Type

Computational parameters.

SoftSol3D.ConfigPar.Source — Type

Source parameters (loading).

SoftSol3D.ConfigPar.SourcePar — Type

Source sub-parameters.

Index

SoftSol3D.ConfigPar

SoftSol3D.ConfigPar.Material

SoftSol3D.ConfigPar.Mesh

SoftSol3D.ConfigPar.Source

SoftSol3D.ConfigPar.SourcePar

SoftSol3D.ConfigPar.initSource

SoftSol3D.LocalMaxSpeeds

SoftSol3D.Run

SoftSol3D.Slice

SoftSol3D.compSol